# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/companies": {
            "get": {
                "description": "Returns a list of companies scraped from screener.in",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Get all companies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CompanyInfo"
                            }
                        }
                    }
                }
            }
        },
        "/api/companies/refresh": {
            "get": {
                "description": "Returns a list of companies scraped from screener.in",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Referesh all companies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CompanyInfo"
                            }
                        }
                    }
                }
            }
        },
        "/api/companies/{name}": {
            "get": {
                "description": "Returns info about a specific company scraped from screener.in",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "companies"
                ],
                "summary": "Get a company details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CompanyInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CompanyInfo": {
            "type": "object",
            "properties": {
                "book_value": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "current_price": {
                    "type": "string"
                },
                "dividend_yield": {
                    "type": "string"
                },
                "face_value": {
                    "type": "string"
                },
                "high_low": {
                    "type": "string"
                },
                "market_cap": {
                    "type": "string"
                },
                "roce": {
                    "type": "string"
                },
                "roe": {
                    "type": "string"
                },
                "stock_pe": {
                    "type": "string"
                }
            }
        }
    }
}